backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
scikit-learn==1.6.1
mediapipe==0.10.32
numpy==1.26.4
opencv-python==4.9.0.80
pydantic==2.5.3
python-multipart==0.0.6
supabase==2.3.0
groq==0.4.1
python-dotenv==1.0.0